requirements.txt
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coverage==6.3.2
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
dnspython==2.2.0
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.29.1
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
openai==0.14.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandas-stubs==1.2.0.49
Pillow==9.0.1
plotly==5.6.0
pluggy==1.0.0
py==1.11.0
pymongo==4.0.1
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
sec-api==1.0.9
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
tenacity==8.0.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.62.3
tzdata==2021.5
urllib3==1.26.8
validators==0.18.2
Werkzeug==2.0.3
